% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\docType{methods}
\name{adapter}
\alias{adapter}
\alias{adapter<-}
\alias{adapter,bcbioSmallRnaDataSet-method}
\alias{adapter<-,bcbioSmallRnaDataSet,list-method}
\title{Sample adapter accesor}
\usage{
\S4method{adapter}{bcbioSmallRnaDataSet}(object)

\S4method{adapter}{bcbioSmallRnaDataSet,list}(object) <- value
}
\arguments{
\item{object}{\link{bcbioSmallRnaDataSet} object.}

\item{...}{Additional arguments.}

\item{value}{A \link{data.frame} with adapter removal informaion.}
}
\value{
\link{list} or \link{bcbioSmallRnaDataSet}.
}
\description{
Sample adapter accesor
}
