% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\docType{methods}
\name{bcbio}
\alias{bcbio}
\alias{bcbio<-}
\alias{bcbio,bcbioSmallRnaDataSet-method}
\alias{bcbio<-,bcbioSmallRnaDataSet-method}
\title{bcbioSmallRnaDataSet caller accessors}
\usage{
\S4method{bcbio}{bcbioSmallRnaDataSet}(object, type = "counts")

\S4method{bcbio}{bcbioSmallRnaDataSet}(object, type = "counts") <- value
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{value}{An integer matrix or other object.}

\item{type}{Type of count data to retrieve.}
}
\value{
caller object.
}
\description{
Additional obejcts of interest include:
}
\details{
\itemize{
\item \href{http://bioconductor.org/packages/3.6/bioc/html/isomiRs.html}{isomiRs}
}
}
