% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcbioSmallRna.R
\docType{package}
\name{bcbioSmallRna-package}
\alias{bcbioSmallRna}
\alias{bcbioSmallRna-package}
\title{bcbioSmallRna}
\description{
Quality control and differential expression for bcbio-nextgen small RNA-seq
experiments.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lpantano/bcbiosRnaseq}
  \item Report bugs at \url{https://github.com/lpantano/bcbiosRnaseq/issues}
}

}
\author{
\strong{Maintainer}: Lorena Pantano \email{lorena.pantano@gmail.com}

Authors:
\itemize{
  \item Michael J. Steinbaugh \email{mike@steinbaugh.com}
}

Other contributors:
\itemize{
  \item Harvard Chan Bioinformatics Core \email{bioinformatics@hsph.harvard.edu} [copyright holder, funder]
}

}
