% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{bcbioSmallRnaDataSet-class}
\alias{bcbioSmallRnaDataSet-class}
\alias{bcbioSmallRnaDataSet}
\title{bcbioSmallRnaDataSet}
\description{
S4 class containing small RNA-seq counts and metadata generated by
\href{https://bcbio-nextgen.readthedocs.org/}{bcbio-nextgen}.
}
\details{
\link{bcbioSmallRnaDataSet} is a subclass of
\link[MultiAssayExperiment:MultiAssayExperiment]{MultiAssayExperiment::MultiAssayExperiment} designed to store an RNA-seq
analysis.

The \code{metadata} accessor contains:
\itemize{
\item Sample summary statistics (a.k.a. metrics).
\item Server run paths.
\item R local environment information, including \code{\link[=sessionInfo]{sessionInfo()}}.
}
}
\author{
Lorena Pantano, Michael Steinbaugh
}
