% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\docType{methods}
\name{counts,bcbioSmallRnaDataSet-method}
\alias{counts,bcbioSmallRnaDataSet-method}
\title{Count matrix accessors}
\usage{
\S4method{counts}{bcbioSmallRnaDataSet}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional parameters.}
}
\value{
Counts matrix
}
\description{
By default, \code{\link[=counts]{counts()}} returns the raw counts. Normalized counts, including
transcripts per million (TPM) can be accessed using the \code{normalized}
argument.
}
\author{
Lorena Pantano
}
