% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download}
\alias{download}
\alias{bcbSmallDownload}
\title{Download Dependency File}
\usage{
bcbSmallDownload()
}
\description{
If the required dependency file isn't present, download latest version from
the \href{http://bioinformatics.sph.harvard.edu/}{HBC website}.
}
\details{
File download utility for RMarkdown knit reports.
}
\examples{
bcbSmallDownload()
}
