% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadRun.R
\name{loadSmallRnaRun}
\alias{loadSmallRnaRun}
\title{Load small RNA-seq bcbio-nextgen run}
\usage{
loadSmallRnaRun(upload_dir = "final", interesting_groups = "description",
  max_samples = 50, ...)
}
\arguments{
\item{upload_dir}{Path to final upload directory. This path is set when
running \code{bcbio_nextgen -w template}.}

\item{interesting_groups}{Character vector of interesting groups. First entry
is used for plot colors during quality control (QC) analysis. Entire vector
is used for PCA and heatmap QC functions.}

\item{max_samples}{\emph{Optional}. Maximum number of samples to calculate rlog and
variance stabilization object from DESeq2.}

\item{...}{Additional arguments, saved as metadata.}
}
\value{
\link{bcbioSmallRnaDataSet}.
}
\description{
Simply point to the final upload directory output by
\href{https://bcbio-nextgen.readthedocs.io/}{bcbio-nextgen}, and this function
will take care of the rest. It automatically imports small RNA-seq counts,
metadata, and program versions used.
}
\note{
When working in RStudio, we recommend connecting to the bcbio-nextgen
run directory as a remote connection over
\href{https://github.com/osxfuse/osxfuse/wiki/SSHFS}{sshfs}.
}
\author{
Michael Steinbaugh, Lorena Pantano
}
