% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\docType{methods}
\name{metrics}
\alias{metrics}
\alias{metrics,bcbioSmallRnaDataSet-method}
\title{Sample metrics accesor}
\usage{
\S4method{metrics}{bcbioSmallRnaDataSet}(object)
}
\arguments{
\item{object}{\link{bcbioSmallRnaDataSet} object.}
}
\description{
Sample metrics accesor
}
