% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\docType{methods}
\name{mirna}
\alias{mirna}
\alias{cluster}
\alias{trna}
\alias{isomir}
\alias{mirna,bcbioSmallRnaDataSet-method}
\alias{cluster}
\alias{trna}
\alias{isomir}
\alias{isomir,bcbioSmallRnaDataSet-method}
\alias{cluster,bcbioSmallRnaDataSet-method}
\alias{trna,bcbioSmallRnaDataSet-method}
\title{mirna count data information}
\usage{
\S4method{mirna}{bcbioSmallRnaDataSet}(object, type = "raw")

\S4method{isomir}{bcbioSmallRnaDataSet}(object, type = "raw")

\S4method{cluster}{bcbioSmallRnaDataSet}(object, type = "raw")

\S4method{trna}{bcbioSmallRnaDataSet}(object, type = "raw")
}
\arguments{
\item{object}{\link{bcbioSmallRnaDataSet} object.}

\item{type}{Type of count data to retrieve.}
}
\description{
mirna count data information
}
