% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-smallrna.R
\name{bcbSmallPCA}
\alias{bcbSmallPCA}
\title{Plot PCA of the different small RNA data}
\usage{
bcbSmallPCA(bcb, type = "mirna", minAverage = 5, columns = NULL,
  data = FALSE, ...)
}
\arguments{
\item{bcb}{\link{bcbioSmallRnaDataSet}.}

\item{type}{Data type to plot: \code{mirna, cluster, isomir, trna}.}

\item{minAverage}{Minimun average small RNA expression to be kept. (log2 scale.)}

\item{columns}{Columns in colData to add to heatmap.}

\item{data}{Only return data.}

\item{...}{Options pass to \code{\link[DEGreport:degPCA]{DEGreport::degPCA()}}.}
}
\description{
Clustering small RNA samples.
}
\author{
Lorena Pantano
}
\keyword{plot}
