% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{sbcb}
\alias{sbcb}
\title{Tiny example of bcbioSmallRna object}
\format{An object of class \code{bcbioSmallRnaDataSet} of length 4.}
\usage{
sbcb
}
\description{
Raw fastq file were downloaded from ftp.sra.ebi.ac.uk.
}
\details{
Data was analyzed with \href{https://bcbio-nextgen.readthedocs.io/en/latest/}{bcbio_nextgen}.
Config files to reproduce the analysis are at:

https://github.com/bcbio/bcbio_srnaseq_output_example

R code to load the data was:\preformatted{sbcb = loadSmallRnaRun("~/repos/bcbio_srnaseq_output_example/final",
interestingGroups = "country")
}

`
}
\author{
Lorena Pantano
}
\keyword{datasets}
