% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-smallrna.R
\name{bcbSmallSize}
\alias{bcbSmallSize}
\alias{bcbSmallMicro}
\alias{bcbSmallCluster}
\alias{bcbSmallSizeDist}
\title{Small RNA-seq quality control plots}
\usage{
bcbSmallSize(bcb, color = NULL)

bcbSmallSizeDist(bcb, color = NULL, percentage = TRUE)

bcbSmallMicro(bcb, color = NULL)

bcbSmallCluster(bcb, color = NULL)
}
\arguments{
\item{bcb}{\link{bcbioSmallRnaDataSet}.}

\item{color}{Column in metadata to use to color the bars.}

\item{percentage}{Whether to plot the percentage or absolute number.}
}
\value{
\link{ggplot}.
}
\description{
Plot size distribution of small RNA-seq data.
}
\examples{
library(DEGreport)
data(sbcb)
bcbSmallSize(sbcb, color = "population")
bcbSmallSizeDist(sbcb, color = "population")
bcbSmallMicro(sbcb, color = "population")
bcbSmallCluster(sbcb, color = "population")
data <- bcbSmallPCA(sbcb, minAverage=8)
degPCA(data[["counts"]], data[["annotation"]], condition = "population")
data <- bcbSmallPCA(sbcb, type = "cluster", minAverage = 8)
degPCA(data[["counts"]], data[["annotation"]], condition = "population")
}
\author{
Lorena Pantano, Michael Steinbaugh
}
\keyword{plot}
