% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-smallrna.R
\name{bcbSmallSize}
\alias{bcbSmallSize}
\alias{bcbSmallMicro}
\alias{bcbSmallCluster}
\alias{bcbSmallSizeDist}
\title{Small RNA-seq quality control plots}
\usage{
bcbSmallSize(bcb, color = NULL)

bcbSmallSizeDist(bcb, color = "NULL")

bcbSmallMicro(bcb, group = NULL)

bcbSmallCluster(bcb)
}
\arguments{
\item{bcb}{\link{bcbioSmallRnaDataSet}.}

\item{color}{Column in metadata to use to color the bars.}
}
\value{
\link{ggplot}.
}
\description{
Plot size distribution of small RNA-seq data.
}
\author{
Lorena Pantano, Michael Steinbaugh
}
\keyword{plot}
